+++ src/components/graph/nodes/question-node.tsx
@@ -18,10 +18,10 @@ enum NodeType {
 
 type QuestionNodeProps = NodeProps & {
   data: {
-    text?: string;
+    text: string;
+    dynamicGeneration: boolean;
     type: NodeType;
     ratings?: string[];
-    dynamicGeneration?: boolean;
   };
 };
 
@@ -64,64 +64,62 @@ export default function QuestionNode({ id, data }: QuestionNodeProps) {
 
   return (
     <NodeContainer title="Question" faIcon={faQuestionCircle} id={id}>
-      <div className="space-y-4 my-4">
-        <Input
-          label="Dynamic Generation"
-          name={`dynamicGeneration_${id}`}
-          type="checkbox"
-          checked={data.dynamicGeneration}
-          onChange={(e) => handleUpdate({ dynamicGeneration: (e.target as HTMLInputElement).checked })}
-          className="nodrag"
-        />
+      <Input
+        label="Dynamic Generation"
+        name="dynamicGeneration"
+        type="checkbox"
+        checked={data.dynamicGeneration}
+        onChange={(e) => handleUpdate({ dynamicGeneration: (e.target as HTMLInputElement).checked })}
+        className="nodrag"
+      />
 
-        <Select
-          label="Question Type"
-          name="type"
-          value={data.type}
-          onChange={(e) => handleUpdate({ type: e.target.value as NodeType })}
-          options={Object.values(NodeType).map((value) => ({ value, label: value }))}
-          className="nodrag"
-        />
+      <Select
+        label="Question Type"
+        name="type"
+        value={data.type}
+        onChange={(e) => handleUpdate({ type: e.target.value as NodeType })}
+        options={Object.values(NodeType).map((value) => ({ value, label: value }))}
+        className="nodrag"
+      />
 
-        <Textarea
-          label="Message"
-          subLabel={
-            data.dynamicGeneration ? 'Prompt for an AI to generate a question' : 'Question asked directly to the user'
-          }
-          name="text"
-          value={data.text}
-          onChange={(e) => handleUpdate({ text: e.target.value })}
-          placeholder="Enter your text here..."
-          className="resize-none overflow-hidden nodrag"
-          rows={3}
-        />
+      <Textarea
+        label="Message"
+        subLabel={
+          data.dynamicGeneration ? 'Prompt for an AI to generate a question' : 'Question asked directly to the user'
+        }
+        name="text"
+        value={data.text}
+        onChange={(e) => handleUpdate({ text: e.target.value })}
+        placeholder="Enter your text here..."
+        className="resize-none overflow-hidden nodrag"
+        rows={3}
+      />
 
-        {!data.dynamicGeneration && (data.type === NodeType.RatingSingle || data.type === NodeType.RatingMulti) && (
-          <div className="flex flex-col gap-4">
-            <label className="text-sm font-medium text-slate-700">Ratings</label>
-            {data.ratings?.map((rating, index) => (
-              <div key={index} className="flex items-center justify-between gap-2">
-                <Input
-                  name={`rating-${index}`}
-                  value={rating}
-                  onChange={(e) => handleRatingChange(index, e.target.value)}
-                  className="flex-grow nodrag"
-                />
-                <Button
-                  onClick={() => removeRating(index)}
-                  variant="transparentSecondary"
-                  size="small"
-                  icon={faTimesCircle}
-                  className="nodrag flex-shrink-0 text-slate-400"
-                />
-              </div>
-            ))}
-            <Button onClick={addRating} variant="primary" size="medium" icon={faPlus} className="nodrag mt-2">
-              Add Rating
-            </Button>
-          </div>
-        )}
-      </div>
+      {!data.dynamicGeneration && (data.type === NodeType.RatingSingle || data.type === NodeType.RatingMulti) && (
+        <div className="flex flex-col gap-4">
+          <label className="text-sm font-medium text-slate-700 dark:text-slate-300">Ratings</label>
+          {data.ratings?.map((rating, index) => (
+            <div key={index} className="flex items-center justify-between gap-2">
+              <Input
+                name={`rating-${index}`}
+                value={rating}
+                onChange={(e) => handleRatingChange(index, e.target.value)}
+                className="flex-grow nodrag"
+              />
+              <Button
+                onClick={() => removeRating(index)}
+                variant="transparentSecondary"
+                size="small"
+                icon={faTimesCircle}
+                className="nodrag flex-shrink-0 text-slate-400"
+              />
+            </div>
+          ))}
+          <Button onClick={addRating} variant="primary" size="medium" icon={faPlus} className="nodrag mt-2">
+            Add Rating
+          </Button>
+        </div>
+      )}
       <Handle type="target" position={Position.Left} className="w-4 h-4 !bg-green-500" />
       <CustomHandle connectionCount={1} type="source" position={Position.Right} className="w-4 h-4 !bg-green-500" />
     </NodeContainer>
